influx/requirements.txt
influxdb_client==1.28.0
paho-mqtt==1.6.1